Move from bullet to prosopite for N+1 query protection [DEV-201] (#6386)
https://github.com/charkost/prosopite

Initial motivation: `prosopite` is more starred in my
[`github-stars-explorer.ts`](https://github.com/davidrunger/dotfiles/blob/cc98273c0a18bcbb6b32d20abca15e8317cf460c/tools/github-stars-explorer.ts).

After trying it out (i.e. this PR), I can understand why it is more
highly starred. The error messages seem better, and it seems to detect
more N+1 queries. Also, IIRC, `bullet` errors immediately on any N+1
query, whereas I think `prosopite` will collect multiple errors across a
request, so that they can all be fixed at once.

One downside of `prosopite` is that we also have to add `pg_query` to
the Gemfile, which in turn requires `google-protobuf`, which seems to
take a while to compile its native extensions. But usually that will be
cached, so not a huge deal.