Add MarriageMembership model (not foreign keys on marriage) [DEV-208] (#6387)
Motivation: this allows us to define User#marriage as a proper Rails
association (which we can eager load), rather than defining it as an
instance method on User (which we cannot eager load). This will allow us
to get rid (in a future PR) of hacks such as [this][1].

[1]:
https://github.com/davidrunger/david_runger/blob/f99333e0b247f83dd4f23a6739f47e122b890662/app/models/user.rb#L57-L69

I think that this data model is also a little bit more elegant, robust,
and versatile. It feels right. As one example, it allows us to more
confidently/accurately track when the second partner joined the marriage
(which currently we could only guess to maybe/probably be
Marriage#updated_at). It's also more flexible; for example, this new
model can accommodate polygamous marriages, which our previous model
could not.

To do this as safely as possible, I will do it in two PRs. The first
will be this PR, which includes a datamigration to copy marriages to the
new join table model. Then, in a subsequent PR, we will drop the
`partner_1_id` and `partner_2_id` columns from the `marriages` table and
remove any code related to those dropped foreign key columns.