[ci] Enable interpolation in rspec command heredocs
I noticed that the `#{rspec_output_options}` didn't seem to be getting
interpolated into the commands being modified in this change, based on
the fact that the test output didn't seem to have `--format failures
--format progress` formatting. That's because the heredoc markers had
single quotes around them, meaning that there is no interpolation. This
change removes the single quotes so that we have interpolation again.

The reason that we had the single-quoted heredoc markers, in the first
place, was so that the `\n` characters wouldn't need to be double
escaped. But we don't actually need those bits of the command, anyway,
so this change simply removes them.