Render flash messages for home page via vue-toastification [DEV-214] (#6435)
**Motivation:** HTML-rendered flash messages are at least
partially/largely overwritten when a Vue app renders into the page, so
it looks pretty bad. This way, it won't look so bad.

This change also updates toast messages for the logs app to be rendered
by `vue-toastification` rather than by `toastify`. This is for
standardization and because the `vue-toastification` toasts have a
butter appearance and a better UX.

The file size increases (JS and CSS) for the home app and the logs app
are not terrible.

## Before


![image](https://github.com/user-attachments/assets/dd3adfc9-85cb-4f3e-80de-18f0d6593112)

## After


![image](https://github.com/user-attachments/assets/7450f044-1ad1-47bf-bf34-b9bd5525ec6e)