Run 'RunTsd' when diff mention tsc and/or typescript
I saw that we are doing a major version bump in `vue-tsc` in this PR and
it makes me feel that we should run `RunTsd`. I don't think it's
relevant, since I think that `vue-tsc` is pretty Vue-specific?, but I'll
feel more confident running it, to be on the safe side. I was thinking
about running it manually; let's just have CI run it, instead.