Don't validate ModelGraphController#index bootstrap data
Since it mostly comes from a gem and then is consumed by a JavaScript
library, it doesn't really seem appropriate or necessary/useful to
validate the shape of the data ourselves. The shape of this data could
change over time, and I don't think we want to have to update our schema
if/when that happens.

I was also motivated to make this change because viewing the model graph
while signed in (thereby putting `current_user` in the bootstrap data)
caused a schema violation. Not worth it.