Specify actions for [skip_]before_actions in ModelGraphController (#6475)
I think that this is a best practice and try to follow it. If we add
another action to this controller, then we might not want these
skip_before_action(s)/before_action(s) to apply to those routes by
default. Since our default before_actions are generally designed for
security, any alterations to them generally makes things less secure,
and so I want to need to opt-in to such alterations.