Sidestep RuboCop 'Style/RequireOrder' warnings by using newlines (#6484)
This is a little cleaner/simpler. Although it certainly does not do so
as clearly as disabling the RuboCop cop, I think that adding the extra
newlines herein does still somewhat hints at the intent to do the first
`require`s before the latter ones. On net, I think this way is better.