[api/logs#destroy] Fix N+1 query (#6491)
Prosopite flagged the N+1 query when I deleted a log locally to test a
change for https://davidrunger.atlassian.net/browse/DEV-217 .

I wonder if `bullet` would also have caught this? It would seem that it
didn't, but we didn't have test coverage for it before, and it's
possible that maybe I just never deleted a log in development since
adding `bullet`. That seems slightly unlikely, though. I'm guessing that
`bullet` had a chance to catch this, but failed. If so, then, nice job,
Prosopite, catching what Bullet missed.