Replace most 'setTimeout' calls with Vue's 'nextTick' [DEV-217] (#6493)
The relevant Vue docs https://vuejs.org/api/general#nexttick don't
explicitly contrast `setTimeout` with `nextTick`, but Perplexity had
this to say:

> For Vue.js applications, nextTick is generally the preferred choice
when you need to perform actions after DOM updates, as it's more
reliable and efficient within the Vue ecosystem.

-- https://www.perplexity.ai/search/0c9c8fe9-5cd6-473e-acbb-9fd4b66cc1c7

So, this change replaces `setTimeout` usage in Vue components (and, in
one case, in a Pinia store) in cases where `setTimeout` was being used
to wait a tick (rather than to actually wait for a certain amount of
time) to instead use `nextTick`.

One small downside of `nextTick` is that it needs to be imported,
whereas `setTimeout` is globally available. Also, it will presumably
cost us at least a tiny bit of page weight, though I imagine it's fairly
negligible.