Only delete relevant TS file if a JSON Schema file deleted [DEV-219]
Prior to this change, we would delete all of the TypeScript types and
then regenerate them. This is pretty inefficient and slow, and I don't
really see a reason to do it this way (apart from it being simple).

Instead, with this change, we'll only delete the specific TypeScript
type file corresponding to the JSON Schema file that has been deleted.