Only delete relevant TS file if a JSON Schema file deleted [DEV-219] (#6494)
Prior to this change, if a JSON Schema file were to be deleted, then we
would delete all of the TypeScript types and then regenerate them all.
This is pretty inefficient and slow, and I don't really see a reason to
do it this way (apart from it being simple).

Instead, with this change, we'll only delete the specific TypeScript
type file corresponding to the JSON Schema file that has been deleted.
This is nearly instantaneous, versus taking quite a while (40 seconds or
something? not sure exactly) to regenerate all of the remaining types,
as we were doing before, which would only have gotten worse as we add
more schemas over time.