Add missing keys to v-for loops where reasonably possible
There are several places where we don't have a key available for the
objects being iterated over that is guaranteed to be unique and also we
cannot construct such a key from the object's properties. In these
cases, I will just continue to omit the `key`, with the thinking that,
while it might hurt performance, at least it will hopefully avoid bugs
that would be caused by mutating and/or non-unique keys.