Sanitize HTML returned by (herein renamed) 'linkifyAndSanitize'
Since the HTML returned by the function formerly known as `linkify` and
now known as `linkifyAndSanitize` will often be dangerously inserted
into the DOM using `v-html`, sanitizing all of the linkified HTML by
default provides a layer of protection against XSS attacks.

In fact, if it weren't for our Content Security Policy (CSP), we would
have had an XSS vulnerability here! Thank goodness that we do have a
pretty strong CSP that provides broad protection against XSS attacks.
But it's good to have multiple, redundant layers of protection, rather
than relying on any single layer of protection, which is why I think
that this change is very much worth making, regardless of the fact that
it's not patching an actual XSS vulnerability.