Rename linkifyAndSanitize => linkifiedAndSanitizedHtml (#6503)
I try to name things that are used for their return value as nouns, and
things that are used for their side effects as verbs. In that paradigm,
since the function being modified herein is used for its return value,
it should be named as a noun, which this change does.