Remove hard wrapping from and generally widen Vue templates [HOME-21] (#6520)
One good thing about doing this is that it will create standardization
in the codebase, whereas currently we have a mix of hard wrapped and not
hard wrapped text blocks.

Another advantage is that we could add the change to
`.git-blame-ignore-revs`.

Another little advantage is that eslint-plugin-vue-pug struggles to
understand this syntax:

``` p. Text down here. ```

and raises some false positives, as a result.