Pluralize the path from 'workout' to 'workouts' [WORK-5]
Since "workouts" (plural) is a part of the functionality of this app
(tracking one's workouts over time, and viewing the workouts of others),
I think it makes more sense to name the path using that plural. Even
when one is doing a workout (i.e. focusing on a single workout), it's
still "within the context" of plural workouts; we are creating a workout
that will be added to the collection of workouts, and it is one of the
many _workouts_ that one might do. This is basically why it makes sense
for Rails controller names to be plural, even though they have "show"
actions (and others that act on only one record at a time). For this
reason, I think we should make "workouts" the path name, which this
change does.