Render Google SSO button/form as a web component (not PNG) [DEV-220] (#6522)
I saw a somewhat blurry sign-in button on my iPhone when not signed in
and looking at the comments section of a blog article.

I think that rendering the button as a web component (with actual HTML
"Sign in with Google" text and an SVG Google icon, rather than a PNG
image of that text and Google icon) is a nice way to deliver the button
along with its associated styles, for use in both Rails views and Vue
components.

Unfortunately, the testing situation with custom elements / shadow DOM
seems to be pretty bleak. I added the `capybara-shadowdom` gem, but it
doesn't work really smoothly, and I sort of had to hack a solution in
the `click_sign_in_with_google` spec helper method that I'm adding in
this change.