[ApplicationPolicy] Require a user for create? and new? (#6537)
We usually want to know who has created the relevant record and who owns
it.

There are some cases where we now need to pass an
AuthToken-authenticated user, which is accomplished via the new
`allow_auth_token_authorization` controller class method.

There are some exceptions to this, which are reflected by the changes
herein, where, for some specific policies (EventPolicy, UserPolicy,
WebhookEmailForwardPolicy), we continue with `created?` always returning
`true` (even if there is no user).