[AuthTokens] Allow scoping to specific action(s) [DEV-222]
Users will have to enter strings like `api/csp_results#create`, which is
a not great UI at all, but it's fine for now, considering that I will be
the only person ever to use this feature.

Limiting a token in this way will significantly decrease the extent of
the security loss if a user's token is ever compromised.