Move tailwindcss from devDependencies to dependencies
Ultimately, it doesn't really matter. We install both sets of packages,
and neither set of packages gets shipped as-is to production; the only
thing that gets shipped to production are the compiled assets. But we
had `@tailwindcss/vite` (in `dependencies`) and `tailwindcss` (in
`devDependencies`) in different groups before, which doesn't seem right.
This change consolidates them into `dependencies` by moving
`tailwindcss` there.

It's not totally clear whether Tailwind should be in devDependencies
(since it includes build tooling) or dependencies (since some of its
code gets built into our compiled assets). I am going with dependencies
because that is what the docs here
https://tailwindcss.com/docs/installation/using-vite implicitly suggest
with the command `npm install tailwindcss @tailwindcss/vite`.