Update SaveRequest spec to reflect HTML escaping
It's good to have a spec that checks that this escaping is working.

Regarding sanitizing the params of `IpBlock`s: on the one hand, it might
be sort of annoying not to have the raw reasons saved, but, on the other
hand, it might be good to guard against potential XSS attacks,
especially in this case an attack that might compromise an AdminUser's
account.