Upload prerenders to 'david-runger-test-uploads' bucket [DEV-228]
Prior to this change, we uploaded prerenders to the
`david-runger-uploads` bucket.

However, I don't think that really makes sense. Why not upload them to
`david-runger-test-uploads`?

`david-runger-test-uploads` is publicly readable, so even though we set
up our S3 / AWS stuff in production (in which environment the prerenders
will need to be downloaded) to use credentials for the
`david-runger-uploads-crud-user` AWS IAM user (which has access to the
`david-runger-uploads` S3 bucket), our S3 client in production should
also still be able to read prerenders that were uploaded to
`david-runger-test-uploads` (since, again, that bucket is publicly
readable).

It should be fine to upload prerenders to a publicly readable bucket.
They contain no secret info. Anyone can ultimately access the prerenders
just by requesting https://davidrunger.com/ .

Therefore, this change switches from uploading prerenders to
`david-runger-uploads` and instead now uploads them to
`david-runger-test-uploads`.

This change has the advantage of not needing to give any step in our CI
process access to the production assets that are stored in the
`david-runger-uploads bucket`.

Additionally, this change simplifies things by allowing us just to use a
single set of AWS credentials in CI, making [this change][1] unnecessary
(hence it is reverted in this PR) and allowing us to delete those
secrets from the GitHub Actions secrets.