Don't HTML-escape special characters with Loofah [DEV-229]
According to the `loofah-activerecord` documentation, this makes our
strings unsafe to render as HTML. But at least the strings should still
be safer than before, when we were making no server-side attempt to
escape HTML at all, since Loofah will still strip HTML tags etc for us.

In the relatively rare cases where we actually do render strings as
HTML, we should be protected by our CSP and also by using DOMPurify, so,
to the extent that an attack can be smuggled in via unsafe characters
(even with all of Loofah's other protections), hopefully we will still
be safe.