Monkeypatch 'Warning' in test so any warnings will fail the test
I sort of thought that there might be a way built into Ruby to do this,
but apparently not (
https://www.perplexity.ai/search/in-ruby-is-there-a-way-to-make-h1aHCodvQ2eUsxbkXcjDtQ
, https://chatgpt.com/c/67e57e66-cb18-8007-bf82-99cdb88064c8 ), so I am
monkeypatching it.

This will ensure that any warnings are caught by CI, rather than being
logged without anyone noticing (and fixing them).

Initially, my monkeypatch simply raised the warning message as an
exception, but that didn't play well with how Cuprite/Ferrum handle
exceptions when the warning is written by cuprite. It caused the spec to
hang indefinitely (using no CPU). So, I switched to adding the
warning(s) to an array and checking it in an `around` block.