[comments] Confirm with user before deleting a comment [COM-7] (#6571)
Although it's sort of annoying to confirm, the user should not be at
risk of losing a comment just because of a single misclick.

Also, monkeypatch `Warning` so that any warnings will fail specs.
**Motivation:** When I didn't wrap the `click_on('Delete')` call in the
spec in an `accept_confirm do ... end` block, the spec still passed, but
it printed a warning. I don't want my code to generate warnings, and I
want to be aware if it does. Failing a spec if it generates any warnings
will help to accomplish those goals.