Only enable dark mode SSO button if given 'dark-mode-supported' class
Otherwise, with just the prior commit in this PR, if a user has dark
mode enabled, then they'd see a dark mode Google SSO button even on
pages that don't have a general dark mode theme (e.g. the main login
page), which looks bad. With this change, in order to provide dark mode
support for the Google login button (presumably because it's in a page
with broader dark mode support), we most provide the
`dark-mode-supported` class to the `google-sign-in-button` custom
element.