[comments] Render dark Google SSO button if user prefers [COM-8] (#6573)
It doesn't look amazing, and just on pure aesthetics I think it looks a
bit worse than a light mode SSO button. For this reason, I was going to
not go ahead with this. However, a downside of the light mode SSO button
is that it stands out sort of a lot and it is kind of a visual
distraction. It can draw one's eye away from the article text as one
gets near the end of the article. That's why I am going ahead with this
change.

| Before | After |
| - | - |
|
![image](https://github.com/user-attachments/assets/eec902f9-97b4-49a0-9166-da803c022d29)
|
![image](https://github.com/user-attachments/assets/29babfa5-6c79-41ac-af79-38248c307426)
|