Bump activeadmin from 3.3.0 to 4.0.0.beta15 [DEV-218] (#6585)
I have no particular motivation for this change, other than to try to
stay up to date with the latest and (hopefully) greatest. This will
reduce the number of gems flagged by `bundle outdated`, which I am happy
about. In fact, with this change, I think that `ferrum` (held back by
`cuprite`) will be our only outdated gem that we can't upgrade.

| Before | After |
| - | - |
|
![image](https://github.com/user-attachments/assets/3bda17c2-3c57-4d51-be8f-780e97e31e1d)
|
![image](https://github.com/user-attachments/assets/d4b8fa1a-bcd1-49a5-8609-046b139ff7eb)
|