Remove support for auth tokens in query params [DEV-236]
Instead, the auth token must be provided via a body param or via a
header.

**Motivation:** Since URLs are not encrypted, including query params in
the URL is way too insecure and should never have been allowed. It could
give access to any auth-token-authenticatable endpoints via a user's
account to any actor able to intercept the URL.

I'm tempted to remove support for params altogether (even body params),
and require that the auth token be provided via a header, but I guess
it's fine. Although I sort of feel like a header is likely to be more
secure, I can't really prove that or even be very confident of it.