[comments] Load styles.css so that we can use Tailwind [COM-9] (#6599)
It feels like sort of a weird quirk (and it's also sort of annoying / a
lost opportunity) that, prior to this change, I could not use Tailwind
classes to style components in the comments app. This change fixes that
by injecting the styles.css style sheet into blog #show pages (the only
ones where we currently render the comments app). This will add some
unnecessary page weight, but I guess that's true of every page where we
load the Tailwind styles, so it's not a new/unique problem, and
styles.css is currently compiling to just 20.2 kB uncompressed and 6.3
kB gzipped, which is not so much, and also I believe that CSS is not as
costly to the browser to process as JavaScript, compared byte-for-byte.