[auth-tokens] Display 'last_used_at' to user [DEV-237] (#6606)
The time will be shown as a relative time, with the ISO-8601 timestamp
available by hovering over it (via a `title` attribute).

This will help users to determine tokens that they aren't using anymore,
and which therefore can safely be deleted, which deletion of
no-longer-needed keys will increase the security of the user's account.