Lower RAILS_MAX_THREADS from 2 to 1 except for web
**Motivation:** Even with our 2 GiB of memory on Hetzner, swap usage
continues to be a bit of a problem. I'm hoping that this change
(decreasing the RAILS_MAX_THREADS from 2 to 1 for the `worker` service)
will decrease memory usage by the `worker` dyno to a helpful extent.

Move PROMETHEUS_EXPORTER_HOST to Dockerfile so that we don't have to
worry about overwriting it (problematic) or redefining it (annoying) in
the `web` service when we override the `environment` section.