Lower RAILS_MAX_THREADS from 2 to 1 except for web (#6608)
**Motivation:** Even with our 2 GiB of memory on Hetzner, swap usage
continues to be a bit of a problem. I'm hoping that this change
(decreasing the RAILS_MAX_THREADS from 2 to 1 for the `worker` service)
will decrease memory usage by the `worker` dyno to a helpful extent.
Really low Sidekiq latency is not important for the app, so lowering the
Sidekiq thread count to 1 should not be a problem.

As part of this change, move `PROMETHEUS_EXPORTER_HOST` to the
`Dockerfile`, so that we don't have to worry about overwriting it
(problematic) or redefining it (annoying) in the `web` service when we
override the `environment` section.