Move /.well-known/traffic-advice response from Rails routes to NGINX (#6610)
I guess that, if we are going to be defining a static route like this,
we might as well do it at the outermost layer of our stack. I guess
there's not much point having it clutter up the Rails routes. I did sort
of like having a test for it, but I guess that's not super necessary.
Since this route will probably change almost never, manually testing it
once (now) should be fine.

https://gemini.google.com/app/38357f101675de23 (poor)

https://chatgpt.com/c/67e98fb6-d89c-8007-9fa3-b4e0103b920c (useful)