[BlogController] Give 'sub!' replacement via block [BLOG-29] (#6613)
This avoids quite the little gotcha when a user's name includes a single
quote:


https://rubyapi.org/3.4/o/string#class-String-label-Substitution+Methods
:

> Note that within the string `replacement` [...] you may refer to some
special match variables using these combinations: [...] `\'` corresponds
to `$'`, which contains the string after the match.

We did not intend to have this behavior. The result was that a lot of
the HTML would be copied into the `html` variable twice (e.g. it would
have two closing `</html>` tags), and in general it messed up the HTML
content.

Providing the replacement string instead as the return value of a block
avoids this issue.