Add a universal logged-in header [DEV-17]
Indicating to the user that they are logged in and giving them the
ability to log out is a basic UX minimum. Users should also be able to
access more information about their account and the ability to take
actions on it (such as deleting their account).

For this reason, this change adds a header that will appear for logged
in users on all pages. The header shows the user's email and also has a
"My account" link and a "Sign out" button.

The header looks and works fine on mobile.

It was a bit of work to get it looking decent on all of our various apps
(in both light mode and dark mode when relevant, such as on the blog),
but I think that it does.

This change removes the logged-in menu info from the sidebar of the
Groceries app, since that is now redundant with this header.

I am going with a header rather than a footer because I think that a
header is more mobile friendly. On a phone, I think that the screen
nearer to the bottom of the screen is more valuable to save for user
interactions, since it is easier to reach with one's thumb. The higher
parts of the screen are less valuable, since they are harder to reach.
Therefore, I am putting the relatively low-interaction logged-in info
etc into a header rather than a footer.