Disable prerendering if a user is signed in [HOME-22]
Currently, no logged-in header appears on the home page, even if a user
is logged in. I believe that this is because of prerendering. To avoid
this issue (and other present or future issues that might be caused by
differences between the HTML that would be served to a logged-in user vs
the prerendered HTML that is generated without a logged-in user), this
change modifies the Prerenderable concern to not use a prerender if a
user is signed in.