[logs] Add some padding beneath the keyboard shortcut tip [LOG-40] (#6635)
This change fixes a visual bug: there was no padding beneath the
keyboard shortcut tip in the logs app. This bug was only noticeable on
short screens and/or with many logs. This change fixes the bug by
changing the margin bottom beneath the tip to a padding bottom. The
margin bottom was having no effect when the tip was the last element on
the page.