[logs/mobile] Don't autofocus input for graph logs [LOG-43]
Bug: When viewing a log with a graph (i.e. not text) on mobile, the
whole graph does not initially appear in the viewport

Cause: This is because of autofocusing the input. This causes the
keyboard to open up (decreasing the screen size) and also for the page
to scroll to focus the input (scrolling some of the graph out of the
view).

Solution: As a solution, we will not autofocus the input when viewing a
log with a graph on a mobile device.