[logs] Center form inputs for non-text logs [LOG-44] (#6642)
Having the text log form be the same as the non-text log form is sort of
rough, since they have pretty different styling and even different form
elements. But it's probably better than duplicating a bunch of logic
code between them. But the best of all might be to extract a composable
with the shared logic, while allowing the templates and styling to be
separate.

| Before | After |
| - | - |
|
![image](https://github.com/user-attachments/assets/3bfd8146-3133-42cb-8744-8ed6136b0e29)
|
![image](https://github.com/user-attachments/assets/1ddcedf0-74ef-42c4-b15e-c35c7df820c8)
|