[specs/proposal] Delete spec that is redundant with a check-ins spec (#6667)
This change deletes a spec from `spec/features/proposal_spec.rb` that
was mostly a redundant duplicate of a spec in
`spec/features/check_ins_spec.rb`, to which I am adding a few lines that
were present in the `spec/features/proposal_spec.rb` spec but not in the
`spec/features/check_ins_spec.rb` spec. I am choosing to preserve the
`spec/features/check_ins_spec.rb` spec because it has a section that
adds an emotional need, which the `spec/features/proposal_spec.rb` spec
did not.