[checkins] Create emotional needs upon marriage creation [CHK-1] (#6670)
This will provide a more useful and convenient starting point for users,
rather than needing to come up with and enter a list of emotional needs
themselves. Users may delete/change/add emotional needs, as they see
fit, but I think a starting point will be useful, as an example, if
nothing else.