[JsonSchemaValidator] Print schema on validation error [LOG-47] (#6676)
We had a spec flake with this error:

```
          JSON::Schema::ValidationError:
            The property '#/definitions/Root/properties/log_id/type' of type string did not match one or more of the required schemas
          # ./app/poros/json_schema_validator.rb:36:in 'JsonSchemaValidator#schema_validation_errors'
```

That validation error is _not_ forced/simulated/reproduced by changing
this
https://github.com/davidrunger/david_runger/blob/7614d554a63dc2d018241c28eb026ef9b9a57e93/spec/support/schemas/api/log_entries/index.json#L23
from `integer` to `string`, but it is forced/simulated/reproduced by
changing it to `integerz`.

I am pretty much at a loss to explain what would have caused this
flakiness, but, to increase our chances of being able to diagnose such
flakiness, in the future, this change will print the schema in the event
of such a validation error.

Relatedly, this change switches from providing a file path to the schema
to instead provide the file's schema content directly. This ensures
that, in the event that the schema is invalid, the schema that we print
is indeed the schema that was used (and found to be invalid) by the
`json-schema` gem.

This change won't actually fix the mysterious-to-me cause of the spec
flakiness that is the subject of the linked ticket. However, I will let
this change close that ticket, since hopefully this will at least
significantly improve our hopes of diagnosing such flakiness, should it
ever occur again in the future.